Added local DB and renamed SchoolItems -> Disciplines
